+++ analysis/filter/StructureModifier/src/main/kotlin/de/maibornwolff/codecharta/filter/structuremodifier/StructureModifierAction.kt
@@ -3,7 +3,7 @@ package de.maibornwolff.codecharta.filter.structuremodifier
 enum class StructureModifierAction(
     val descripton: String
 ) {
-    PRINT_STRUCTURE("Print the structure of the project"),
+    PRINT_STRUCTURE("Print the structure of the project (deprecated, please use the 'inspect' command instead)"),
     SET_ROOT("Extract a sub path as the new root"),
     MOVE_NODES("Move nodes within the project"),
     REMOVE_NODES("Remove nodes"),